requirements.txt
# Core dependencies
opencv-python>=4.8.0
pyautogui>=0.9.54
mss>=9.0.0
numpy>=1.24.0
Pillow>=10.0.0
customtkinter>=5.2.0
pynput>=1.7.7
pyinstaller>=6.0.0

# Platform-specific dependencies
# Linux
python-xlib>=0.33; sys_platform == 'linux'

# macOS
pyobjc-framework-Quartz>=9.0; sys_platform == 'darwin'

# Windows
pywin32>=305; sys_platform == 'win32'